requirements.txt
Flask~=2.2.2
selenium~=4.4.3
python-dotenv~=0.21.0